% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bi_chain_cpp}
\alias{bi_chain_cpp}
\title{A Gibbs sampler using Rcpp}
\usage{
bi_chain_cpp(num = 10000L, a = 1L, b = 1L, n = 100L)
}
\arguments{
\item{num}{The number of samples. Default is \code{10000}.}

\item{a}{Parameter of density. Default is \code{1}.}

\item{b}{Parameter of density. Default is \code{1}.}

\item{n}{Parameter of density. Default is \code{100}.}
}
\value{
A \code{matrix}, the number of rows is the param num and the numberof colmuns is 2.
}
\description{
A Gibbs sampler using Rcpp, the target density has been given.
}
\examples{
\dontrun{
c.chain <- bi_chain_cpp()
plot(c.chain)
}
}
