% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.fit.R
\name{fit.gamma.appro}
\alias{fit.gamma.appro}
\title{Fit the gamma model with approximate newton method.}
\usage{
fit.gamma.appro(x, y, weight = NULL, start = NULL, lambda = 0)
}
\arguments{
\item{x}{Input matrix, of dimension \eqn{n \times p}; each row is an observation
vector and each column is a predictor/feature/variable.}

\item{y}{The response variable, of \code{n} observations.}

\item{weight}{Observation weights. When \code{weight = NULL},
we set \code{weight = 1} for each observation as default.}

\item{start}{the start point of fitness.}

\item{lambda}{A single lambda value for L2 regularized. Default is 0.}
}
\value{
Intercept and linear coefficient.
}
\description{
Use approximate newton method fit the gamma model which is an example of generalized linear model.
}
\examples{
\dontrun{
n <- 1000
p <- 10
dataset <- generate.data(n, p, p,family = "gamma", seed = 1)
coef <- fit.gamma.appro(dataset[["x"]],dataset[["y"]])
}
}
