% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_fit_approximate_newton_method}
\alias{gamma_fit_approximate_newton_method}
\title{gamma fit approximate newton method}
\usage{
gamma_fit_approximate_newton_method(x, y, weights, beta, coef0, lambda)
}
\arguments{
\item{x}{Input matrix, of dimension \eqn{n \times p}; each row is an observation vector and each column is a predictor/feature/variable.}

\item{y}{The response variable, of \code{n} observations.}

\item{weights}{Observation weights.}

\item{beta}{Linear coefficient.}

\item{coef0}{Intercept.}

\item{lambda}{L2 penalty coefficient.}
}
\description{
fit gamma model with approximate newton method.
}
