% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_fit_IWLS_method}
\alias{gamma_fit_IWLS_method}
\title{gamma fit IWLS method}
\usage{
gamma_fit_IWLS_method(x, y, weights, beta, coef0, lambda)
}
\arguments{
\item{x}{Input matrix, of dimension \eqn{n \times p}; each row is an observation vector and each column is a predictor/feature/variable.}

\item{y}{The response variable, of \code{n} observations.}

\item{weights}{Observation weights.}

\item{beta}{Initial value of linear coefficient.}

\item{coef0}{Initial value of intercept.}

\item{lambda}{L2 penalty coefficient.}
}
\value{
Intercept and linear coefficient.
}
\description{
fit gamma model with IWLS method.
}
